change name of package to lattice_errors
